requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-allauth==0.51.0
django-crispy-forms==1.14.0
django-extensions==3.2.0
django-tinymce
djongo==1.3.6
idna==3.3
jsmin==3.0.1
mysql-connector-python==8.0.29
oauthlib==3.2.0
Pillow==9.2.0
pkg_resources==0.0.0
protobuf==4.21.2
psycopg2-binary==2.9.3
pycparser==2.21
pyenchant==3.2.2
PyJWT==2.4.0
pymongo==4.1.1
python3-openid==3.2.0
requests==2.28.1
requests-oauthlib==1.3.1
sqlparse==0.2.4
urllib3==1.26.10
